Revert "Merge pull request #66 from MilleniumStudio/wod_foundry"
This reverts commit 9710ea27f895e34e466872edeea4c11a57b44f2c, reversing
changes made to fce4d07c28c900c204299000462b0c596d0ecd97.

+++ src/server/game/Achievements/AchievementMgr.cpp
@@ -273,7 +273,7 @@ bool AchievementCriteriaData::IsValid(CriteriaEntry const* p_Criteria)
     }
 }
 
-bool AchievementCriteriaData::Meets(uint32 p_CriteriaID, Player const* p_Source, Unit const* p_Target, uint64 p_MiscValue1 /*= 0*/) const
+bool AchievementCriteriaData::Meets(uint32 p_CriteriaID, Player const* p_Source, Unit const* p_Target, uint32 p_MiscValue1 /*= 0*/) const
 {
     switch (dataType)
     {
@@ -386,7 +386,7 @@ bool AchievementCriteriaData::Meets(uint32 p_CriteriaID, Player const* p_Source,
     return false;
 }
 
-bool AchievementCriteriaDataSet::Meets(Player const* p_Source, Unit const* p_Target, uint64 p_MiscValue /*= 0*/) const
+bool AchievementCriteriaDataSet::Meets(Player const* p_Source, Unit const* p_Target, uint32 p_MiscValue /*= 0*/) const
 {
     for (Storage::const_iterator itr = storage.begin(); itr != storage.end(); ++itr)
         if (!itr->Meets(criteria_id, p_Source, p_Target, p_MiscValue))
